move constants/messages into content/
