+++ client/components/CourseInfo.jsx
@@ -3,7 +3,7 @@ import React from "react";
 import { useIntl } from "react-intl";
 
 import { campuses, departments, terms, years } from "../constants";
-import { messages } from "../constants/intl/components/CourseInfo";
+import { messages } from "../content/messages/components/CourseInfo";
 
 const CourseInfo = ({
   errors,
