+++ client/pages/admin/index.jsx
@@ -8,8 +8,8 @@ import CheckPermissions from "../../components/CheckPermissions";
 import RequestsList from "../../components/RequestsList";
 import SectionContainer from "../../components/SectionContainer";
 import UsersList from "../../components/UsersList";
-import { messages } from "../../constants/intl/pages/admin";
 import { REQUEST_RESULT } from "../../constants/toasts";
+import { messages } from "../../content/messages/pages/admin";
 import { mapAsOption } from "../../helpers";
 import { getAdminData } from "../../requests";
 import { modifyGroupchatStatus } from "../../requests/groupChats";
